Matrix multiplication  by Vector ready
